// src/student/interfaces/student.interface.ts
export interface StudentInterface {
    studentId?: string
    refId?: string
    aadhaar?: string
    firstName?: string
    lastName?: string
    schoolId?: string
    currentClassId?: string
    gender?: string
    socialCategory?: string
    iscwsn?: string
    religion?: string
    singleGirl?: string
    weight?: string
    height?: string
    bloodGroup?: string
    birthDate?: string
    homeless?: string
    bpl?: string
    migrant?: string
    status?: string
    email?: string
  }